+++ tests/testthat/test-fitVARMxID-fitvarmxid-nu-theta-mu0.R
@@ -75,20 +75,48 @@ lapply(
       robust = FALSE,
       seed = 42
     )
+    sav <- FitVARMxIDSave(
+      data = data,
+      observed = paste0("y", seq_len(k)),
+      id = "id",
+      center = FALSE,
+      alpha_fixed = TRUE,
+      alpha_values = alpha,
+      psi_fixed = TRUE,
+      psi_d_values = psi_ldl$uc_d,
+      psi_l_values = psi_ldl$s_l,
+      nu_fixed = FALSE,
+      theta_fixed = FALSE,
+      theta_diag = FALSE,
+      theta_d_values = rep(x = 1, times = k),
+      mu0_fixed = FALSE,
+      mu0_values = mu0,
+      robust = FALSE,
+      seed = 42
+    )
     if (ci) {
       print(fit)
       print(fit, means = TRUE)
       summary(fit)
       summary(fit, means = TRUE)
-      summary(fit, var_metric = "logvar")
-      summary(fit, var_metric = "softplusvar")
       coef(fit)
-      coef(fit, var_metric = "logvar")
-      coef(fit, var_metric = "softplusvar")
       vcov(fit)
-      vcov(fit, var_metric = "logvar")
-      vcov(fit, var_metric = "softplusvar")
       vcov(fit, robust = TRUE)
+      confint(fit)
+      confint(fit, robust = TRUE)
+      plot(fit)
+      plot(fit, robust = TRUE)
+      print(sav)
+      print(sav, means = TRUE)
+      summary(sav)
+      summary(sav, means = TRUE)
+      coef(sav)
+      vcov(sav)
+      vcov(sav, robust = TRUE)
+      confint(sav)
+      confint(sav, robust = TRUE)
+      plot(sav)
+      plot(sav, robust = TRUE)
     }
     testthat::test_that(
       paste(text, "converged"),