+++ NEWS.md
@@ -1,3 +1,10 @@
+# fitVARMxID 1.0.5
+
+## Patch
+
+* Added `confint()` and `plot()` methods.
+* Added the `FitVARMxIDSave()` function.
+
 # fitVARMxID 1.0.3
 
 ## Patch